[hotfix-51590][gbase]support merge
