+++ flinkx-connectors/flinkx-connector-gbase/src/main/java/com/dtstack/flinkx/connector/gBase/converter/GBaseRawTypeConverter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.dtstack.flinkx.connector.gbase.converter;
+package com.dtstack.flinkx.connector.gBase.converter;
 
 import com.dtstack.flinkx.throwable.UnsupportedTypeException;
